+++ app/src/main/java/app/revanced/manager/ui/screen/InstalledAppInfoScreen.kt
@@ -385,8 +385,14 @@ fun InstalledAppInfoScreen(
             .padding(horizontal = 16.dp)
             .clip(RoundedCornerShape(24.dp))
     ) {
-        val installType = installedApp.installType
         val hasRoot = viewModel.rootInstaller.hasRootAccess()
+        val installType = remember(installedApp.installType, viewModel.primaryInstallerIsMount, hasRoot) {
+            if (installedApp.installType == InstallType.SAVED && viewModel.primaryInstallerIsMount && hasRoot) {
+                InstallType.MOUNT
+            } else {
+                installedApp.installType
+            }
+        }
         val rootRequiredText = stringResource(R.string.installer_status_requires_root)
 
         if (viewModel.appInfo != null) {